compilation fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@16329 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Dropbox/src/dropbox_utils.cpp
@@ -80,7 +80,7 @@ void CDropbox::SendToContact(MCONTACT hContact, const char* data)
 {
 	if (hContact == GetDefaultContact()) {
 		char *message = mir_utf8encode(data);
-		AddEventToDb(hContact, EVENTTYPE_MESSAGE, DBEF_UTF, mir_strlen(message), (PBYTE)message);
+		AddEventToDb(hContact, EVENTTYPE_MESSAGE, DBEF_UTF, (DWORD)mir_strlen(message), (PBYTE)message);
 		return;
 	}
 
@@ -100,7 +100,7 @@ void CDropbox::SendToContact(MCONTACT hContact, const char* data)
 
 	if (CallContactService(hContact, PSS_MESSAGE, 0, (LPARAM)data) != ACKRESULT_FAILED) {
 		char *message = mir_utf8encode(data);
-		AddEventToDb(hContact, EVENTTYPE_MESSAGE, DBEF_UTF | DBEF_SENT, mir_strlen(message), (PBYTE)message);
+		AddEventToDb(hContact, EVENTTYPE_MESSAGE, DBEF_UTF | DBEF_SENT, (DWORD)mir_strlen(message), (PBYTE)message);
 	}
 }
